fix:人口統計グラフを取得する非同期処理関数を外部ファイル化、PrefetureList.tsxをmemo化することでレンダリングを最適化
