+++ src/components/List/PrefectureSelect.tsx
@@ -1,0 +1,6 @@
+import React, { FC } from 'react'
+import styles from '../../styles/Home.module.css'
+
+export const prefectureSelect: FC = () => {
+  return <footer className={styles.footer}>Author by Taishin</footer>
+}
