+++ src/components/List/PrefecturesList.tsx
@@ -1,29 +1,18 @@
-import React, { FC, useEffect, useState } from 'react'
+import React, { FC, useEffect, useState, memo } from 'react'
 import { prefecture } from '../../types/prefecture'
 import styles from '../../styles/components/prefectureList.module.css'
-import { APIURL } from '../../common/const'
-import axios from 'axios'
 import { Chart } from '../Chart/Chart'
+import { getChart } from '../../lib/fetch'
 
 type Props = {
   prefectures: prefecture[]
 }
 
-type Headers = {
-  'X-API-KEY': any
-}
-
-// TODO useMemo, useCallbackが必要なコンポーネント
-export const PrefecturesList: FC<Props> = (props) => {
+export const PrefecturesList: FC<Props> = memo((props) => {
   const { prefectures } = props
   const [prefectureCode, setPrefectureCode] = useState<number>(1)
   const [populationData, setPopulationData] = useState<{}>({})
   const [prefectureName, setPrefectureName] = useState<string>('北海道')
-
-  // prefectureCodeが変更されたら、非同期処理を実行する
-  const headers: Headers = {
-    'X-API-KEY': process.env.NEXT_PUBLIC_API_KEY,
-  }
 
   const getPrefCodeAndName = (prefCode: number, prefName: string) => {
     setPrefectureCode(prefCode)
@@ -31,17 +20,10 @@ export const PrefecturesList: FC<Props> = (props) => {
   }
 
   useEffect(() => {
-    const fetcher = async () => {
-      const response = await axios.get(
-        `${APIURL}/api/v1/population/composition/perYear?cityCode=-&prefCode=${prefectureCode}`,
-        {
-          headers,
-        }
-      )
-      setPopulationData(response.data.result.data[0].data)
-      console.log(response.data)
-    }
-    fetcher()
+    ;(async () => {
+      const fetcherData = await getChart(prefectureCode)
+      setPopulationData(fetcherData)
+    })()
   }, [prefectureCode])
 
   return (
@@ -68,4 +50,4 @@ export const PrefecturesList: FC<Props> = (props) => {
       </div>
     </>
   )
-}
+})
